fix: adjust ongoing for small devices and reduce ongoing screens files to one
