+++ Fiero/Views/Challenges/Quick/QCChallengeCreatedView.swift
@@ -21,8 +21,7 @@ struct QCChallengeCreatedView: View {
     @EnvironmentObject var quickChallengeViewModel: QuickChallengeViewModel
     
     @State var didPushToHomeScreen: Bool = false
-    @State var presentDuelChallenge: Bool = false
-    @State var present3Or4Challenge: Bool = false
+    @State var isPresentingChallenge: Bool = false
     @State var isPresentingAlert: Bool = false
     @State var subscriptions: Set<AnyCancellable> = []
     
@@ -71,12 +70,7 @@ struct QCChallengeCreatedView: View {
                             case .failure(_):
                                 self.isPresentingAlert.toggle()
                             case .finished:
-                                if self.quickChallenge.maxTeams == 2 {
-                                    self.presentDuelChallenge.toggle()
-                                }
-                                else {
-                                    self.present3Or4Challenge.toggle()
-                                }
+                                self.isPresentingChallenge.toggle()
                         }
                     }, receiveValue: { _ in ()})
                     .store(in: &subscriptions)
@@ -94,13 +88,8 @@ struct QCChallengeCreatedView: View {
             })
             .padding(.bottom, Tokens.Spacing.xxxs.value)
             
-            NavigationLink("", isActive: self.$presentDuelChallenge) {
-                DuelScreenView(quickChallenge: $quickChallenge, isShowingAlertOnDetailsScreen: self.$isPresentingAlert)
-            }
-            .hidden()
-            
-            NavigationLink("", isActive: self.$present3Or4Challenge) {
-                Ongoing3Or4WithPauseScreenView(quickChallenge: self.$quickChallenge, isShowingAlertOnDetailsScreen: self.$isPresentingAlert)
+            NavigationLink("", isActive: self.$isPresentingChallenge) {
+                OngoingWithPause(quickChallenge: self.$quickChallenge, isShowingAlertOnDetailsScreen: self.$isPresentingAlert)
             }
             .hidden()
         }
